+++ src/app/index/mod.rs
@@ -1,13 +1,10 @@
-use anyhow::*;
 use diesel;
-use diesel::prelude::*;
 use log::error;
 use std::sync::{Arc, Condvar, Mutex};
-use std::time;
+use std::time::Duration;
 
-use crate::app::vfs;
-use crate::config::MiscSettings;
-use crate::db::{misc_settings, DB};
+use crate::app::{config, vfs};
+use crate::db::DB;
 
 mod metadata;
 mod query;
@@ -24,14 +21,16 @@ pub use self::update::*;
 pub struct Index {
 	db: DB,
 	vfs_manager: vfs::Manager,
+	config_manager: config::Manager,
 	pending_reindex: Arc<(Mutex<bool>, Condvar)>,
 }
 
 impl Index {
-	pub fn new(db: DB, vfs_manager: vfs::Manager) -> Self {
+	pub fn new(db: DB, vfs_manager: vfs::Manager, config_manager: config::Manager) -> Self {
 		let index = Self {
 			db,
 			vfs_manager,
+			config_manager,
 			pending_reindex: Arc::new((Mutex::new(false), Condvar::new())),
 		};
 
@@ -76,21 +75,14 @@ impl Index {
 	fn automatic_reindex(&self) {
 		loop {
 			self.trigger_reindex();
-			let sleep_duration = {
-				let connection = self.db.connect();
-				connection
-					.and_then(|c| {
-						misc_settings::table
-							.get_result(&c)
-							.map_err(|e| Error::new(e))
-					})
-					.map(|s: MiscSettings| s.index_sleep_duration_seconds)
-					.unwrap_or_else(|e| {
-						error!("Could not retrieve index sleep duration: {}", e);
-						1800
-					})
-			};
-			std::thread::sleep(time::Duration::from_secs(sleep_duration as u64));
+			let sleep_duration = self
+				.config_manager
+				.get_index_sleep_duration()
+				.unwrap_or_else(|e| {
+					error!("Could not retrieve index sleep duration: {}", e);
+					Duration::from_secs(1800)
+				});
+			std::thread::sleep(sleep_duration);
 		}
 	}
 }
