+++ src/app/user/manager.rs
@@ -1,7 +1,8 @@
-use anyhow::*;
+use anyhow::anyhow;
 use diesel;
 use diesel::prelude::*;
 
+use super::*;
 use crate::db::DB;
 
 const HASH_ITERATIONS: u32 = 10000;
@@ -16,7 +17,35 @@ impl Manager {
 		Self { db }
 	}
 
-	pub fn auth(&self, username: &str, password: &str) -> Result<bool> {
+	pub fn create_user(&self, username: &str, password: &str) -> Result<(), Error> {
+		if password.is_empty() {
+			return Err(Error::EmptyPassword);
+		}
+		let password_hash = hash_password(password)?;
+		let connection = self.db.connect()?;
+		let new_user = User {
+			name: username.to_owned(),
+			password_hash,
+			admin: 0,
+		};
+		diesel::insert_into(users::table)
+			.values(&new_user)
+			.execute(&connection)
+			.map_err(|_| Error::Unspecified)?;
+		Ok(())
+	}
+
+	pub fn set_password(&self, username: &str, password: &str) -> Result<(), Error> {
+		let password_hash = hash_password(password)?;
+		let connection = self.db.connect()?;
+		diesel::update(users::table.filter(users::name.eq(username)))
+			.set(users::password_hash.eq(password_hash))
+			.execute(&connection)
+			.map_err(|_| Error::Unspecified)?;
+		Ok(())
+	}
+
+	pub fn auth(&self, username: &str, password: &str) -> anyhow::Result<bool> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		match users
@@ -33,14 +62,14 @@ impl Manager {
 		}
 	}
 
-	pub fn count(&self) -> Result<i64> {
+	pub fn count(&self) -> anyhow::Result<i64> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		let count = users.count().get_result(&connection)?;
 		Ok(count)
 	}
 
-	pub fn exists(&self, username: &str) -> Result<bool> {
+	pub fn exists(&self, username: &str) -> anyhow::Result<bool> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		let results: Vec<String> = users
@@ -50,7 +79,7 @@ impl Manager {
 		Ok(results.len() > 0)
 	}
 
-	pub fn is_admin(&self, username: &str) -> Result<bool> {
+	pub fn is_admin(&self, username: &str) -> anyhow::Result<bool> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		let is_admin: i32 = users
@@ -60,7 +89,12 @@ impl Manager {
 		Ok(is_admin != 0)
 	}
 
-	pub fn lastfm_link(&self, username: &str, lastfm_login: &str, session_key: &str) -> Result<()> {
+	pub fn lastfm_link(
+		&self,
+		username: &str,
+		lastfm_login: &str,
+		session_key: &str,
+	) -> anyhow::Result<()> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		diesel::update(users.filter(name.eq(username)))
@@ -72,7 +106,7 @@ impl Manager {
 		Ok(())
 	}
 
-	pub fn get_lastfm_session_key(&self, username: &str) -> Result<String> {
+	pub fn get_lastfm_session_key(&self, username: &str) -> anyhow::Result<String> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		let token = users
@@ -89,7 +123,7 @@ impl Manager {
 		self.get_lastfm_session_key(username).is_ok()
 	}
 
-	pub fn lastfm_unlink(&self, username: &str) -> Result<()> {
+	pub fn lastfm_unlink(&self, username: &str) -> anyhow::Result<()> {
 		use crate::db::users::dsl::*;
 		let connection = self.db.connect()?;
 		diesel::update(users.filter(name.eq(username)))
@@ -99,7 +133,7 @@ impl Manager {
 	}
 }
 
-pub fn hash_password(password: &str) -> Result<String> {
+fn hash_password(password: &str) -> anyhow::Result<String> {
 	match pbkdf2::pbkdf2_simple(password, HASH_ITERATIONS) {
 		Ok(hash) => Ok(hash),
 		Err(e) => Err(e.into()),
