+++ src/service/rocket/mod.rs
@@ -28,6 +28,8 @@ pub fn get_server(context: service::Context) -> Result<rocket::Rocket> {
 	Ok(rocket::custom(config)
 		.manage(context.db)
 		.manage(context.index)
+		.manage(context.config_manager)
+		.manage(context.lastfm_manager)
 		.manage(context.playlist_manager)
 		.manage(context.thumbnail_manager)
 		.manage(context.user_manager)
