+++ src/service/test/protocol.rs
@@ -2,8 +2,7 @@ use http::{method::Method, Request};
 use percent_encoding::{percent_encode, NON_ALPHANUMERIC};
 use std::path::Path;
 
-use crate::app::user;
-use crate::config;
+use crate::app::{config, user};
 use crate::service::dto;
 
 pub struct RequestBuilder {}
