+++ src/service/test/settings.rs
@@ -1,6 +1,6 @@
 use http::StatusCode;
 
-use crate::config;
+use crate::app::config;
 use crate::service::test::{constants::*, ServiceType, TestService};
 use crate::test_name;
